fix: arruma base da publicacao
