+++ frontend/src/components/Publication/styles.css
