+++ frontend/src/components/PublicationAntigaa/styles.css
@@ -1,4 +1,4 @@
-/* .publication-content {
+.publication-content {
     width: 90%;
     background-color: #373737;
     padding: 16px;
@@ -32,6 +32,10 @@
 
 .button-movie-search {
     background: #9d9097;
+}
+
+.interactive-into h1 {
+    font-size: 14px;
 }
 
 .button-cancel {
@@ -154,6 +158,11 @@
     width: 69%;
 }
 
+.review-full{
+    text-align: justify;
+    word-break: break-word;
+}
+
 .movie-options {
     position: absolute;
     z-index: 10000;
@@ -178,6 +187,10 @@
 
 .input-search-movie-selected {
     background-color: #6e6e6e;
+}
+
+.like-button{
+    font-size: 20px;
 }
 
 @keyframes spin {
@@ -295,6 +308,253 @@ select {
     animation: spin 2s linear infinite;
 }
 
+.movie-indice{
+    display: flex;
+    font-size: 14px;
+    padding: 8px;
+    padding-bottom: 15px;
+}
+
+.movie-insert{
+    padding-left: 5px;
+    padding-right: 5px;
+    color: white;
+    cursor: pointer;
+    transition: 0.2s;
+}
+
+.movie-insert:hover{
+    color: #ff0381;
+    transition: 0.2s;
+}
+
+.critic-review-preview{
+    display: block;
+}
+
+.critic-review{
+    display: block;
+}
+
+.read-more{
+    padding: 5px;
+    color: white;
+    cursor: pointer;
+    text-align: right;
+    transition: 0.2s;
+}
+
+.read-more:hover{
+    color: #ff0381;
+    transition: 0.2s;
+}
+
+.arrow-display{
+    padding-left: 5px;
+    margin-bottom: -3px;
+}
+
+.user-indice{
+    display: flex;
+}
+
+.user-insert{
+    padding-left: 5px;
+    color: white;
+    cursor: pointer;
+    transition: 0.2s;
+}
+
+.user-insert:hover{
+    color: #ff0381;
+    transition: 0.2s;
+}
+
+.user-insert-comment{
+    padding-left: 5px;
+    color: white;
+    cursor: pointer;
+    transition: 0.2s;
+    display:flex;
+}
+
+.user-insert-comment:hover{
+    color: #ff0381;
+    transition: 0.2s;
+}
+
+.date-release{
+    font-size: 13px;
+    color: rgba(255, 255, 255, 0.562);
+    margin-top: 3px;
+    margin-left: 5px;
+}
+
+.date-release{
+    font-size: 13px;
+    color: rgba(255, 255, 255, 0.562);
+    margin-top: 3px;
+    margin-left: 5px;
+}
+
+
+.image-review{
+    padding-top: 10px;
+    border-radius: 8px;
+    width: 100%;
+}
+
+.rating-content{
+    padding-top: 15px;
+    padding-bottom: 15px;
+    cursor: pointer;
+    float: left;
+}
+
+.star-critic{
+    font-size: 24px;
+    display: flex;
+}
+
+div .pub-star{
+    color: #ff0381b5;
+}
+
+div .no-star{
+    color: #ffffff;
+    cursor: pointer;
+}
+
+.rating-text{
+    font-size: 14px;
+    color: #ffffff;
+    transition: color 0.2s ease-in-out;
+}
+
+.rating-text:hover{
+    color: #ff0381b5;
+}
+
+.del-publication{
+    border-radius: 100%;
+    font-size: 20px;
+    position: absolute;
+    margin-left: 43%;
+    padding: 3px;
+    transition: 0.2s;
+    cursor: pointer;
+}
+
+.del-publication:hover{
+    background-color: #494949;
+    transition: 0.2s;
+}
+
+.options-publication-inside{
+    padding: 5px;
+    cursor: pointer;
+    transition: 0.1s;
+}
+
+.options-publication-inside:hover{
+    background-color: #ff0381b5;
+    transition: 0.1s;
+}
+
+.zone-interactive-publication{
+    display: flex;
+    padding-bottom: 10px;
+}
+
+.interactive-into{
+    width: 50%;
+    text-align-last: center;
+    padding: 10px;
+    transition: 0.2s;
+    border-radius: 6px;
+    background-color: #00000000;
+}
+
+.interactive-into:hover{
+    transition: 0.2s;
+    background-color: rgba(110, 110, 110, 0.527);
+    cursor: pointer;
+}
+
+.interactive-into-likes{
+    width: 50%;
+    display: flex;
+}
+
+.comment-send-button{
+    background-color: #00000000;
+}
+
+.zone-interactive-publication-options{
+    display: flex;
+    padding-left: 25px;
+    padding-right: 25px;
+    padding-bottom: 10px;
+}
+
+.comment-send{
+    color: white;
+    padding-left: 10px;
+    padding-right: 10px;
+}
+
+.comment-send-button{
+    padding: 5px;
+    border-radius: 10px;
+    transition: 0.2s;
+}
+
+.comment-send-button:hover{
+    background-color: rgba(110, 110, 110, 0.527);
+    transition: 0.2s;
+}
+
+.comment-send-button-press{
+    transition: 0.2s;
+    background-color: #ff0381b5;
+}
+
+.comment-send-button-press:hover{
+    transition: 0.2s;
+    background-color: #ff1c8db5;
+}
+
+.self-comment-on{
+    display: flex;
+    padding-left: 30px;
+    padding-right: 30px;
+    padding-top: 10px;
+    padding-bottom: 10px;
+    border: solid 1px rgba(255, 255, 255, 0.034);
+    border-radius: 8px;
+}
+
+.self-comment-on:hover{
+    border: solid 1px rgba(255, 255, 255, 0.068);
+}
+
+.read-more-comments{
+    padding: 5px;
+}
+
+.read-more-comments:hover{
+    width: auto;
+    text-decoration: underline;
+    cursor: pointer;
+}
+
+.comment-view{
+    padding-left: 20px;
+    padding-right: 20px;
+    padding-bottom: 10px;
+    padding-top: 10px;
+}
+
 @keyframes spin {
     0% {
         transform: rotate(0deg);
@@ -316,6 +576,39 @@ select {
     justify-content: center;
     align-items: center;
     z-index: 100;
+}
+
+@media only screen and (max-width: 680px) {
+    .image-review{
+        margin-left: -35px;
+    }
+    .review-full{
+        margin-left: -65px;
+    }
+    .del-publication{
+        border-radius: 100%;
+        font-size: 20px;
+        position: absolute;
+        margin-left: 90%;
+        padding: 3px;
+        transition: 0.2s;
+        cursor: pointer;
+    }
+    .comment-view{
+        padding-left: 5px;
+        padding-right: 5px;
+        padding-bottom: 10px;
+        padding-top: 10px;
+    }
+    .self-comment-on{
+        display: flex;
+        padding-left: 35px;
+        padding-right: 35px;
+        padding-top: 10px;
+        padding-bottom: 10px;
+        border: solid 1px rgba(255, 255, 255, 0.034);
+        border-radius: 8px;
+    }
 }
 
 @media only screen and (max-width: 768px) {
@@ -353,4 +646,4 @@ select {
         height: auto;
     }
 
-} */
+}
