+++ frontend/src/pages/publication/index.js
@@ -1,107 +1,107 @@
-import React, { useEffect, useState } from "react";
-import './styles.css';
+// import React, { useEffect, useState } from "react";
+// import './styles.css';
 
-import Header from "../../components/header";
-import HeaderDesktop from "../../components/headerDesktop";
+// import Header from "../../components/header";
+// import HeaderDesktop from "../../components/headerDesktop";
 
-import ViewPublication from '../../components/ViewPublication';
+// import Publication from '../../components/Publication';
 
-import api from "../../api";
+// import api from "../../api";
 
-import { useParams, useLocation } from 'react-router-dom';
+// import { useParams, useLocation } from 'react-router-dom';
 
-import Menu from "../../components/menu";
+// import Menu from "../../components/menu";
 
-import { Link } from "react-router-dom";
+// import { Link } from "react-router-dom";
 
-import { MdArrowBack } from "react-icons/md";
+// import { MdArrowBack } from "react-icons/md";
 
-const Publication = () => {
-    const { id } = useParams();
+// const Publication = () => {
+//     const { id } = useParams();
 
-    const [publication, setPublication] = useState(null)
+//     const [publication, setPublication] = useState(null)
 
-    const location = useLocation();
+//     const location = useLocation();
 
-    const backButtonRoute = location.state?.prevPath || '/';
+//     const backButtonRoute = location.state?.prevPath || '/';
 
-    const [windowSize, setWindowSize] = useState({
-        width: window.innerWidth,
-        height: window.innerHeight
-    });
+//     const [windowSize, setWindowSize] = useState({
+//         width: window.innerWidth,
+//         height: window.innerHeight
+//     });
 
-    useEffect(() => {
-        const handleResize = () => {
-            setWindowSize({
-                width: window.innerWidth,
-                height: window.innerHeight
-            });
-        };
+//     useEffect(() => {
+//         const handleResize = () => {
+//             setWindowSize({
+//                 width: window.innerWidth,
+//                 height: window.innerHeight
+//             });
+//         };
 
-        window.addEventListener('resize', handleResize);
+//         window.addEventListener('resize', handleResize);
 
-        return () => {
-            window.removeEventListener('resize', handleResize);
-        };
-    }, [])
+//         return () => {
+//             window.removeEventListener('resize', handleResize);
+//         };
+//     }, [])
 
-    useEffect(() => {
-        async function get_publication() {
-            const response = await api.get(`/publicacoes/${id}/`)
-            setPublication(response.data)
-        }
+//     useEffect(() => {
+//         async function get_publication() {
+//             const response = await api.get(`/publicacoes/${id}/`)
+//             setPublication(response.data)
+//         }
 
-        get_publication()
-    }, [])
+//         get_publication()
+//     }, [])
 
-    return (
-        <>
-            {(window.innerWidth > 760) ?
-                <HeaderDesktop />
-                :
+//     return (
+//         <>
+//             {(window.innerWidth > 760) ?
+//                 <HeaderDesktop />
+//                 :
 
-                <Header />
-            }
-            <div className="content-home">
-                {windowSize.width < 680
-                    ?
-                    (
-                        <Menu />
-                    )
-                    :
-                    <div className="home-left-content">
-                        <Menu />
-                    </div>
-                }
-                <div className="content-box-home">
-                    <div className="title-content">
-                        <Link
-                            className="back-btn"
-                            to={backButtonRoute}
-                            style={{ textDecoration: "none", color: "#fff" }}
-                        >
-                            <MdArrowBack size={32} className="back-icon" />
-                        </Link>
-                        <h1 className="title-component">Publicação</h1>
-                    </div>
+//                 <Header />
+//             }
+//             <div className="content-home">
+//                 {windowSize.width < 680
+//                     ?
+//                     (
+//                         <Menu />
+//                     )
+//                     :
+//                     <div className="home-left-content">
+//                         <Menu />
+//                     </div>
+//                 }
+//                 <div className="content-box-home">
+//                     <div className="title-content">
+//                         <Link
+//                             className="back-btn"
+//                             to={backButtonRoute}
+//                             style={{ textDecoration: "none", color: "#fff" }}
+//                         >
+//                             <MdArrowBack size={32} className="back-icon" />
+//                         </Link>
+//                         <h1 className="title-component">Publicação</h1>
+//                     </div>
 
-                    {publication !== null && (<ViewPublication
-                        userID={publication.user_id}
-                        idPost={publication?.id}
-                        idMovie={publication.movie_id}
-                        rating={publication.review}
-                        critic={publication.pub_text}
-                        image={publication?.imgur_link}
-                        date={publication.date}
-                        myPub={false}
-                    />)}
-                </div>
-                <div className="home-right-content">
+//                     {publication !== null && (<Publication
+//                         userID={publication.user_id}
+//                         idPost={publication?.id}
+//                         idMovie={publication.movie_id}
+//                         rating={publication.review}
+//                         critic={publication.pub_text}
+//                         image={publication?.imgur_link}
+//                         date={publication.date}
+//                         myPub={false}
+//                     />)}
+//                 </div>
+//                 <div className="home-right-content">
 
-                </div>
-            </div>
-        </>
-    )
-}
+//                 </div>
+//             </div>
+//         </>
+//     )
+// }
 
-export default Publication;
+// export default Publication;
